Show previous day's goals with completion checkbox in edit
Remove left pane display of previous daily (it was too difficult to 
understand in context the way it was shown). What I really want to see
is the previous daily’s goals with the ability to mark them as complete
on the form, and on the show page the view-only version of the previous
daily’s goals.